refactor(29561): relocate components
