+++ hivemq-edge/src/frontend/src/modules/DomainOntology/components/MetadataExplorer.tsx
